Add files via upload
Email Spam Classification is a supervised method of classifying the dataset into predefined lables - spam and ham. It is a machine learning project which uses Bernoulli Naive Bayes classifier to classify the mails to spam and ham. The model predicted the class lables with an accuracy of 98% and predicted correctly on the new data provided as input to it.